+++ front-end/src/tests/bookies/bookies.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, test, beforeEach } from '@jest/globals';
-import { Constants } from './samples/constants';
+import { Constants } from './constants';
 import { Rebuild, FlatParent } from '@/modules/bookies/bookies.js';
 
 let c;
